live update preview from document
