+++ package.json
@@ -106,7 +106,7 @@
     "@types/google__maps": "^0.5.14",
     "@types/graphql-date": "^1.0.5",
     "@types/graphql-type-json": "^0.3.2",
-    "@types/html-to-text": "^5.1.0",
+    "@types/html-to-text": "^9.0.0",
     "@types/htmlparser2": "^3.10.1",
     "@types/intercom-client": "^2.11.10",
     "@types/jest": "^26.0.19",
@@ -212,7 +212,7 @@
     "he": "^1.1.1",
     "history": "^3.0.0",
     "html-lexer": "^0.4.1",
-    "html-to-text": "^2.1.0",
+    "html-to-text": "^9.0.3",
     "htmlparser2": "^3.9.2",
     "import": "0.0.6",
     "intercom-client": "^2.10.2",
