+++ packages/lesswrong/components/editor/EditorFormComponent.tsx
@@ -72,6 +72,19 @@ export const EditorFormComponent = ({form, formType, formProps, document, name, 
       }
     }
   }, [getLocalStorageHandlers, currentEditorType]);
+
+  /**
+   * Update the edited field (e.g. "contents") so that other form components can access the updated value. The direct motivation for this
+   * was for SocialPreviewUpload, which needs to know the body of the post in order to generate a preview description and image.
+   */
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  const throttledSetContentsValue = useCallback(_.throttle(async () => {
+    if (!(editorRef.current && shouldSubmitContents(editorRef.current))) return
+    
+    // Preserve other fields in "contents" which may have been sent from the server
+    updateCurrentValues({[fieldName]: {...(document[fieldName] || {}), ...(await editorRef.current.submitData(null))}})
+  }, autosaveInterval, {leading: true}),
+  [saveBackup, autosaveInterval])
   
   // eslint-disable-next-line react-hooks/exhaustive-deps
   const throttledSaveBackup = useCallback(
@@ -99,11 +112,12 @@ export const EditorFormComponent = ({form, formType, formProps, document, name, 
     // this is a serialization of the whole document which can be too slow to do
     // on every keystroke).
     updateCurrentValues({[`${fieldName}_type`]: change.contents?.type});
+    void throttledSetContentsValue()
     
     if (autosave) {
       throttledSaveBackup(contents);
     }
-  }, [throttledSaveBackup, updateCurrentValues, fieldName, isCollabEditor]);
+  }, [isCollabEditor, updateCurrentValues, fieldName, throttledSetContentsValue, throttledSaveBackup]);
   
   useEffect(() => {
     const unloadEventListener = (ev) => {
