+++ packages/lesswrong/components/posts/PostsPage/PostsPage.tsx
@@ -33,7 +33,7 @@ const POST_DESCRIPTION_EXCLUSIONS: RegExp[] = [
 ];
 
 /** Get a og:description-appropriate description for a post */
-export const getPostDescription = (post: PostsWithNavigation | PostsWithNavigationAndRevision | PostsEditQueryFragment) => {
+export const getPostDescription = (post: {contents?: {plaintextDescription: string | null} | null, shortform: boolean, user: {displayName: string} | null}) => {
   if (post.contents?.plaintextDescription) {
     // concatenate the first few paragraphs together up to some reasonable length
     const plaintextPars = post.contents.plaintextDescription
