+++ packages/lesswrong/lib/collections/posts/schema.tsx
@@ -1304,8 +1304,8 @@ const schema: SchemaType<DbPost> = {
     optional: true,
     label: "Social Preview Image",
     viewableBy: ['guests'],
-    editableBy: ['sunshineRegiment', 'admins'],
-    insertableBy: ['sunshineRegiment', 'admins'],
+    editableBy: [userOwns, 'sunshineRegiment', 'admins'],
+    insertableBy: ['members', 'sunshineRegiment', 'admins'],
     control: "SocialPreviewUpload",
     group: formGroups.socialPreview,
     order: 4,
