+++ packages/lesswrong/server/resolvers/revisionResolvers.ts
@@ -3,13 +3,13 @@ import { dataToMarkdown, dataToHTML, dataToCkEditor } from '../editor/conversion
 import { highlightFromHTML, truncate } from '../../lib/editor/ellipsize';
 import { htmlStartingAtHash } from '../extractHighlights';
 import { augmentFieldsDict } from '../../lib/utils/schemaUtils'
-import { sanitize, sanitizeAllowedTags } from '../vulcan-lib/utils';
 import { defineQuery } from '../utils/serverGraphqlUtil';
-import htmlToText from 'html-to-text'
+import { htmlToText } from 'html-to-text'
 import sanitizeHtml, {IFrame} from 'sanitize-html';
 import { extractTableOfContents } from '../tableOfContents';
 import * as _ from 'underscore';
 import { dataToDraftJS } from './toDraft';
+import { sanitize, sanitizeAllowedTags } from '../../lib/vulcan-lib/utils';
 
 const PLAINTEXT_HTML_TRUNCATION_LENGTH = 4000
 const PLAINTEXT_DESCRIPTION_LENGTH = 2000
@@ -77,8 +77,7 @@ augmentFieldsDict(Revisions, {
       resolver: ({html}) => {
         if (!html) return
         const truncatedHtml = truncate(sanitize(html), PLAINTEXT_HTML_TRUNCATION_LENGTH)
-        return htmlToText
-          .fromString(truncatedHtml, {ignoreHref: true, ignoreImage: true, wordwrap: false })
+        return htmlToText(truncatedHtml, {wordwrap: false, selectors: [ { selector: 'img', format: 'skip' }, { selector: 'a', options: { ignoreHref: true } } ] })
           .substring(0, PLAINTEXT_DESCRIPTION_LENGTH)
       }
     }
@@ -100,8 +99,7 @@ augmentFieldsDict(Revisions, {
           }
         )
         const truncatedHtml = truncate(mainTextHtml, PLAINTEXT_HTML_TRUNCATION_LENGTH)
-        return htmlToText
-          .fromString(truncatedHtml)
+        return htmlToText(truncatedHtml)
           .substring(0, PLAINTEXT_DESCRIPTION_LENGTH)
       }
     }
