+++ packages/lesswrong/server/scripts/fixBodyField.ts
@@ -1,5 +1,5 @@
 import { Posts } from '../../lib/collections/posts';
-import htmlToText from 'html-to-text';
+import { htmlToText } from 'html-to-text';
 import { asyncForeachSequential } from '../../lib/utils/asyncUtils';
 
 const runFix = false;
@@ -12,7 +12,7 @@ if (runFix) { void (async ()=>{
   await asyncForeachSequential(allPosts, async (post) => {
     const { html = "" } = post.contents || {}
     if (html) {
-      const plaintextBody = htmlToText.fromString(html);
+      const plaintextBody = htmlToText(html);
       const excerpt =  plaintextBody.slice(0,140);
       await Posts.rawUpdateOne(post._id, {$set: {body: plaintextBody, excerpt: excerpt}});
       postCount++;
